+++ lib/lamby/sam_helpers.rb
@@ -1,9 +1,0 @@
-module Lamby
-  module SamHelpers
-
-    def sam_local?
-      ENV['AWS_SAM_LOCAL'] == 'true'
-    end
-
-  end
-end